refactor(): optimize file structure and content matching
