+++ packages/components/kawa-Material/image/Image.types.ts
