+++ packages/components/tailwind.config.js
@@ -1,15 +1,11 @@
 /** @type {import('tailwindcss').Config} */
 export default {
   content: [
-    "./index.html",
-    "./src/**/*.{js,ts,jsx,tsx}",
-    "./components/**/*.{js,ts,jsx,tsx}",
-    "./stories/**/*.{js,ts,jsx,tsx}",
-    "./**/*.{js,ts,jsx,tsx}",
-    "!**/node_modules/**" // 排除 node_modules 目录
+    './kawa-Material/**/*.{js,jsx,ts,tsx}',
+    './stories/**/*.{js,jsx,ts,tsx}'
   ],
   theme: {
     extend: {},
   },
   plugins: [],
-}
+};
